[Si443x] Use 32-bit IRQ flags

+++ src/modules/Si443x/Si443x.cpp
@@ -300,7 +300,7 @@ int16_t Si443x::startReceive() {
   return(state);
 }
 
-int16_t Si443x::startReceive(uint32_t timeout, uint16_t irqFlags, uint16_t irqMask, size_t len) {
+int16_t Si443x::startReceive(uint32_t timeout, uint32_t irqFlags, uint32_t irqMask, size_t len) {
   (void)timeout;
   (void)irqFlags;
   (void)irqMask;
